push

+++ ft_check.c
@@ -6,7 +6,7 @@
 /*   By: hakader <hakader@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/26 18:31:33 by hakader           #+#    #+#             */
-/*   Updated: 2024/11/29 17:47:44 by hakader          ###   ########.fr       */
+/*   Updated: 2024/11/30 14:02:41 by hakader          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,9 +24,9 @@ int	ft_check(va_list args, char c)
 	else if (c == 'd' || c == 'i')
 		i = ft_putnbr(va_arg(args, int));
 	else if (c == 'x')
-		i = ft_puthexa(va_arg(args, unsigned long), "0123456789abcdef");
+		i = ft_puthexa(va_arg(args, int), "0123456789abcdef");
 	else if (c == 'X')
-		i = ft_puthexa(va_arg(args, unsigned long), "0123456789ABCDEF");
+		i = ft_puthexa(va_arg(args, int), "0123456789ABCDEF");
 	else if (c == 'p')
 		i = ft_puthexa(va_arg(args, int), "0123456789ABCDEF");
 	else if (c == 'u')
